requirements.txt
python-dotenv
discord.py
cookies-discord-components
requests
geopy
pytz
